chore(playground): separate layouts for pages
