+++ playground/src/layouts/ComponentsLayout.vue
@@ -1,0 +1,60 @@
+<script setup>
+import { computed } from 'vue';
+import { useRoute, RouterView } from 'vue-router';
+import UiApp from '@ui/components/App/Index.vue';
+import RouterLink from '../components/RouterLink.vue';
+import { sideBarItems } from '../sideBarItems';
+
+const route = useRoute();
+
+const pageNavItems = [
+  {
+    label: 'Components',
+    children: [
+      { label: 'Buttons', href: '/components/buttons' },
+      { label: 'Forms', href: '/components/forms' },
+      { label: 'Tables', href: '/components/tables' },
+      { label: 'Tabs', href: '/components/tabs' },
+      { label: 'Overlays', href: '/components/overlays' },
+      { label: 'Editor', href: '/components/editor', parent: '/components/editor' },
+    ],
+  },
+];
+
+const pageTabs = computed(() => {
+  if (route.path.startsWith('/components/forms')) {
+    return [
+      { title: 'General', href: '/components/forms' },
+      { title: 'Sizing', href: '/components/forms/sizing' },
+      { title: 'Errors', href: '/components/forms/errors' },
+    ];
+  }
+  if (route.path.startsWith('/components/editor')) {
+    return [
+      { title: 'Overview', href: '/components/editor' },
+      { title: 'Variants', href: '/components/editor/variant' },
+      { title: 'Text', href: '/components/editor/text' },
+    ];
+  }
+  return [];
+});
+
+const pageTitle = computed(() => route.meta.title || '');
+</script>
+
+<template>
+  <UiApp
+    :pageUrl="route.fullPath"
+    :pageTitle="pageTitle"
+    :sideBarItems="sideBarItems"
+    :pageNavItems="pageNavItems"
+    :pageTabs="pageTabs"
+    :linkComponent="RouterLink"
+    :isSideNav="true"
+  >
+    <template #navLogo>
+      <img src="/atlas.png" alt="Atlas" class="h-8 w-8 rounded-full" />
+    </template>
+    <RouterView />
+  </UiApp>
+</template>
