--- playground/src/pages/components/Index.vue
@@ -1,7 +1,0 @@
-<script setup>
-import { RouterView } from 'vue-router';
-</script>
-
-<template>
-  <RouterView />
-</template>
