+++ playground/src/sideBarItems.js
@@ -1,0 +1,17 @@
+import { IconHome, IconUsers, IconSettings } from '@tabler/icons-vue';
+
+export const sideBarItems = [
+  {
+    children: [
+      { label: 'Home', href: '/', icon: IconHome, activeIcon: IconHome },
+      { label: 'Users', href: '/users', icon: IconUsers, activeIcon: IconUsers },
+      {
+        label: 'Components',
+        href: '/components/buttons',
+        parent: '/components',
+        icon: IconSettings,
+        activeIcon: IconSettings,
+      },
+    ],
+  },
+];
